pyproject.toml
[tool.poetry]
name = "fridaylivekit"
version = "0.1.0"
description = ""
authors = ["satyamdas03 <116027621+satyamdas03@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
livekit-agents = "^1.1.4"
livekit-plugins-openai = "^1.1.4"
livekit-plugins-silero = "^1.1.4"
livekit-plugins-google = "^1.1.4"
livekit-plugins-noise-cancellation = "^0.2.4"
mem0ai = "^0.1.112"
duckduckgo-search = "^8.0.4"
langchain-community = "^0.3.26"
requests = "^2.32.4"
python-dotenv = "^1.1.1"
pypdf = "^5.6.1"
langchain-openai = "^0.3.27"
faiss-cpu = "^1.11.0"
livekit-plugins-deepgram = ">=1.1.4,<1.2.0"
livekit-api = "^1.0.3"
livekit-plugins-spitch = "^1.1.4"
livekit-plugins-sarvam = ">=1.0,<2.0"
aiohttp = "^3.12.13"
fastapi = "^0.115.14"
uvicorn = {extras = ["standard"], version = "^0.35.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
